+++ src/main/java/org/springframework/social/partnercenter/api/PartnerCenterTemplate.java
@@ -9,6 +9,8 @@ import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.social.oauth2.AbstractOAuth2ApiBinding;
 import org.springframework.social.partnercenter.PartnerCenter;
+import org.springframework.social.partnercenter.api.agreement.AgreementMetaDataOperations;
+import org.springframework.social.partnercenter.api.agreement.AgreementMetaDataTemplate;
 import org.springframework.social.partnercenter.api.agreement.AgreementOperations;
 import org.springframework.social.partnercenter.api.agreement.AgreementTemplate;
 import org.springframework.social.partnercenter.api.analytics.AnalyticsOperations;
@@ -75,6 +77,7 @@ public class PartnerCenterTemplate extends AbstractOAuth2ApiBinding implements P
 	private final DirectoryRoleOperations directoryRoleOperations;
 	private final RoleOperations roleOperations;
 	private final ConsentOperations consentOperations;
+	private final AgreementMetaDataOperations agreementMetaDataOperations;
 	private final AgreementOperations agreementOperations;
 
 	public PartnerCenterTemplate(UriProvider uriProvider, String accessToken, String version) {
@@ -125,6 +128,9 @@ public class PartnerCenterTemplate extends AbstractOAuth2ApiBinding implements P
 		roleOperations = new RoleTemplate(createRestResource(uriProvider.partnerBaseUri().build().toUri()), isAuthorized());
 
 		consentOperations = new ConsentTemplate(createRestResource(uriProvider.partnerCenterCustomerUri().build().toUri()), isAuthorized());
+
+		agreementMetaDataOperations = new AgreementMetaDataTemplate(createRestResource(
+				uriProvider.partnerCenterBuilder().pathSegment("v1").build().toUri()), isAuthorized());
 
 		agreementOperations = new AgreementTemplate(createRestResource(
 				uriProvider.partnerCenterBuilder().pathSegment("v1").build().toUri()), isAuthorized());
@@ -229,6 +235,11 @@ public class PartnerCenterTemplate extends AbstractOAuth2ApiBinding implements P
 	@Override
 	public ConsentOperations getConsentOperations() {
 		return consentOperations;
+	}
+
+	@Override
+	public AgreementMetaDataOperations getAgreementMetaDataOperations() {
+		return agreementMetaDataOperations;
 	}
 
 	@Override