+++ src/main/java/org/springframework/social/partnercenter/api/agreement/AgreementMetaDataOperations.java
@@ -1,0 +1,9 @@
+package org.springframework.social.partnercenter.api.agreement;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.social.partnercenter.api.PagingResourceOperations;
+import org.springframework.social.partnercenter.api.PartnerCenterResponse;
+
+public interface AgreementMetaDataOperations extends PagingResourceOperations<AgreementMetaData> {
+	ResponseEntity<PartnerCenterResponse<AgreementMetaData>> getAgreementMetadatas();
+}