+++ src/main/java/org/springframework/social/partnercenter/api/agreement/AgreementOperations.java
@@ -4,8 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.social.partnercenter.api.PagingResourceOperations;
 import org.springframework.social.partnercenter.api.PartnerCenterResponse;
 
-public interface AgreementOperations extends PagingResourceOperations<AgreementMetaData> {
-	ResponseEntity<PartnerCenterResponse<AgreementMetaData>> getAgreementMetadatas();
+public interface AgreementOperations extends PagingResourceOperations<Agreement> {
 	ResponseEntity<Agreement> confirmCustomerAcceptance(String customerTenantId, Agreement agreement);
 	ResponseEntity<PartnerCenterResponse<Agreement>> getConfirmations(String customerTenantId);
 }