+++ src/main/java/org/springframework/social/partnercenter/connect/PartnerCenterMultiTenantConnectionFactory.java
@@ -1,7 +1,5 @@
 package org.springframework.social.partnercenter.connect;
 
-import java.util.Optional;
-
 import org.springframework.social.oauth2.AccessGrant;
 import org.springframework.social.partnercenter.PartnerCenter;
 import org.springframework.social.partnercenter.api.uri.SecurityRegion;
@@ -16,21 +14,25 @@ public class PartnerCenterMultiTenantConnectionFactory extends BasePartnerCenter
     private final String clientId;
     private final String clientSecret;
 
-    public PartnerCenterMultiTenantConnectionFactory(String clientId, String clientSecret){
+    private final SecurityRegion securityRegion;
+
+    public PartnerCenterMultiTenantConnectionFactory(String clientId, String clientSecret, SecurityRegion securityRegion) {
         super(PartnerCenter.PROVIDER_ID, null, new PartnerCenterApiAdapter());
         this.clientId = clientId;
         this.clientSecret = clientSecret;
-        this.multiTenantOAuthOperations = new AzureADMultiTenantOAuthTemplate(clientId, clientSecret, UriProvider.DEFAULT_URL_PROVIDER);
+        this.securityRegion = securityRegion;
+        this.multiTenantOAuthOperations = new AzureADMultiTenantOAuthTemplate(clientId, clientSecret, UriProvider.fromSecurityRegion(securityRegion));
     }
 
     /**
      * Create a connection to Partner Center
-     * @param refreshToken obtained from {@link AzureADMultiTenantOAuthOperations#exchangeForRefreshToken}
+     *
+     * @param refreshToken    obtained from {@link AzureADMultiTenantOAuthOperations#exchangeForRefreshToken}
      * @param partnerTenantId tenantId of the csp account for which the connection is created
-     * @param domain domain of the csp account
-     * @param securityRegion security region of the csp account if null will default to US
+     * @param domain          domain of the csp account
+     * @return the PartnerCenterConnection that allows operations with Partner Center APIs
      */
-    public PartnerCenterConnection createConnection(String refreshToken, String partnerTenantId, String domain, SecurityRegion securityRegion) {
+    public PartnerCenterConnection createConnection(String refreshToken, String partnerTenantId, String domain) {
         AccessGrant accessGrant = multiTenantOAuthOperations
                 .exchangeRefreshTokenForAccess(refreshToken, securityRegion.getPartnerServiceApiRoot(), partnerTenantId);
 
@@ -38,14 +40,13 @@ public class PartnerCenterMultiTenantConnectionFactory extends BasePartnerCenter
                 extractProviderUserId(accessGrant),
                 accessGrant.getAccessToken(),
                 accessGrant.getExpireTime(),
-                new PartnerCenterServiceProvider(clientId, clientSecret, domain, Optional.ofNullable(securityRegion)
-                        .orElse(SecurityRegion.USA)),
+                new PartnerCenterServiceProvider(clientId, clientSecret, domain, securityRegion),
                 getApiAdapter());
     }
 
-    public boolean canConnect(String refreshToken, String partnerTenantId, String domain, SecurityRegion securityRegion) {
+    public boolean canConnect(String refreshToken, String partnerTenantId, String domain) {
         try {
-            createConnection(refreshToken, partnerTenantId, domain, securityRegion).getApi().getCustomerOperations().getList(1);
+            createConnection(refreshToken, partnerTenantId, domain).getApi().getCustomerOperations().getList(1);
             return true;
         } catch (Exception e) {
             return false;
@@ -57,8 +58,12 @@ public class PartnerCenterMultiTenantConnectionFactory extends BasePartnerCenter
     }
 
     @Override
-    public void enableSl4fjForAuthRequests(LogLevel logLevel){
+    public void enableSl4fjForAuthRequests(LogLevel logLevel) {
         super.enableSl4fjForAuthRequests(logLevel);
         multiTenantOAuthOperations.enableSlf4j(logLevel);
+    }
+
+    public SecurityRegion getSecurityRegion() {
+        return securityRegion;
     }
 }