+++ src/main/java/org/springframework/social/partnercenter/security/AzureADMultiTenantOAuthTemplate.java
@@ -4,11 +4,8 @@ import static java.util.Optional.ofNullable;
 import static org.springframework.social.partnercenter.api.uri.UriProvider.DEFAULT_URL_PROVIDER;
 import static org.springframework.social.partnercenter.api.validation.Assertion.notNull;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
 import java.util.ArrayList;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -24,7 +21,6 @@ import org.springframework.social.oauth2.OAuth2Parameters;
 import org.springframework.social.partnercenter.api.ApiAuthorizationException;
 import org.springframework.social.partnercenter.api.AuthorizationFault;
 import org.springframework.social.partnercenter.api.uri.UriProvider;
-import org.springframework.social.partnercenter.connect.PartnerCenterJWT;
 import org.springframework.social.partnercenter.http.logging.HttpRequestResponseLoggerFactory;
 import org.springframework.social.partnercenter.http.logging.LogLevel;
 import org.springframework.social.partnercenter.http.logging.LoggingRequestInterceptor;
@@ -37,8 +33,6 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
 
-import com.google.common.base.Charsets;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 
 public class AzureADMultiTenantOAuthTemplate implements AzureADMultiTenantOAuthOperations {
@@ -86,26 +80,6 @@ public class AzureADMultiTenantOAuthTemplate implements AzureADMultiTenantOAuthO
 						.forEach(entry -> params.put(entry.getKey(), entry.getValue())));
 
 		return postForAccessGrant(uriProvider.buildPartnerCenterOAuth2Uri(partnerTenantId), params).getRefreshToken();
-	}
-
-	@Override
-	public DelegatedAccessGrant extractDelegatedAccessGrant(String body) {
-		final Map<String, String> response = Splitter.on('&').trimResults().withKeyValueSeparator("=").split(body)
-				.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> urlDecode(e.getValue())));
-
-		final String code = response.get("code");
-		final PartnerCenterJWT partnerJWT = PartnerCenterJWT.fromTokenString(response.get("id_token"));
-
-		String state = response.get("state");
-		return new DelegatedAccessGrant(code, state, partnerJWT.getTid());
-	}
-
-	private String urlDecode(String param) {
-		try {
-			return URLDecoder.decode(param, Charsets.UTF_8.toString());
-		} catch (UnsupportedEncodingException e) {
-			throw new RuntimeException(e);
-		}
 	}
 
 	@Override